Refactor: Organize preprocessing into static TextProcessingUtils class
